requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
pandas>=2.2.0
python-decouple==3.8
openpyxl>=3.1.0
xlrd>=2.0.0 
django-storages>=1.14.0
boto3>=1.34.0
drf-spectacular>=0.27.0
djangorestframework-simplejwt>=5.3.1

psycopg2-binary
django-redis>=5.4.0
redis>=5.0.0
polib>=1.2.0 
# Google OAuth
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
requests==2.32.4
beautifulsoup4>=4.12.0
lxml>=4.9.0
PyJWT>=2.10.0
cryptography>=41.0.0
python-dotenv>=1.0.0
openai>=1.0.0
google-generativeai>=0.3.0
channels>=4.0.0
channels-redis>=4.1.0
websocket-client>=1.6.0
PyYAML>=6.0.0
celery>=5.3.0
celery[redis]>=5.3.0 